pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = "A production-ready Django REST Framework API template"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "django-environ>=0.11",
    "django-cors-headers>=4.4",
    "django-celery-beat>=2.6",
    "psycopg[binary]>=3.2",
    "celery[redis]>=5.4",
    "gunicorn>=22",
    "whitenoise>=6.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]

# ── Ruff ──────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
src = ["apps", "config", "tests"]
line-length = 88
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "DJ",   # flake8-django
    "RUF",  # ruff-specific
]
ignore = ["RUF012"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

# ── mypy ──────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = false
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
mypy_path = "."

[[tool.mypy.overrides]]
module = ["tests.*", "apps.*.tests.*", "conftest"]
disallow_untyped_defs = false
disable_error_code = ["call-arg"]

[[tool.mypy.overrides]]
module = ["apps.*"]
disable_error_code = ["misc", "type-arg"]
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = ["config.settings.*"]
disable_error_code = ["misc", "name-defined"]

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "django_celery_beat.*",
    "environ.*",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.dev"

# ── pytest ────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
testpaths = ["tests", "apps"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --reuse-db"

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.1",
    "djangorestframework-stubs[compatible-mypy]>=3.15",
    "factory-boy>=3.3",
    "mypy>=1.13",
    "pre-commit>=4.0",
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "ruff>=0.8",
]
